Album/requirements.txt
filetype==1.2.0
PyExifTool==0.5.6